Revert "1735 - Updated test scripts to reflect splitting imputation into two jobs"
This reverts commit 053a0c7b30ff24b22c71e7c18df08e02e285de3b.
