+++ projects/_03_independent_cqc/_03_impute/tests/fargate/test_impute_ind_cqc_ascwds_and_pir.py
@@ -1,7 +1,17 @@
 import unittest
 from unittest.mock import ANY, Mock, patch
 
+import polars as pl
+import polars.testing as pl_testing
+
 import projects._03_independent_cqc._03_impute.fargate.impute_ind_cqc_ascwds_and_pir as job
+from projects._03_independent_cqc.unittest_data.polars_ind_cqc_test_file_data import (
+    ImputeIndCqcAscwdsAndPirData as Data,
+)
+from projects._03_independent_cqc.unittest_data.polars_ind_cqc_test_file_schemas import (
+    ImputeIndCqcAscwdsAndPirSchema as Schemas,
+)
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 PATCH_PATH = (
     "projects._03_independent_cqc._03_impute.fargate.impute_ind_cqc_ascwds_and_pir"
@@ -17,10 +27,20 @@ class ImputeIndCqcAscwdsAndPirTests(unittest.TestCase):
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")
     @patch(f"{PATCH_PATH}.utils.nullify_ct_values_previous_to_first_submission")
     @patch(f"{PATCH_PATH}.model_imputation")
+    @patch(f"{PATCH_PATH}.merge_ascwds_and_pir_filled_post_submissions")
+    @patch(f"{PATCH_PATH}.convert_pir_to_filled_posts")
+    @patch(f"{PATCH_PATH}.model_primary_service_rate_of_change_trendline")
+    @patch(f"{PATCH_PATH}.cUtils.calculate_filled_posts_per_bed_ratio")
+    @patch(f"{PATCH_PATH}.forward_fill_latest_known_value")
     @patch(f"{PATCH_PATH}.utils.scan_parquet", return_value=mock_data)
     def test_main_runs_successfully(
         self,
         scan_parquet_mock: Mock,
+        forward_fill_latest_known_value_mock: Mock,
+        calculate_filled_posts_per_bed_ratio_mock: Mock,
+        model_roc_trendline_mock: Mock,
+        convert_pir_to_filled_posts_mock: Mock,
+        merge_ascwds_and_pir_filled_post_submissions_mock: Mock,
         model_imputation_mock: Mock,
         nullify_ct_values_previous_to_first_submission_mock: Mock,
         sink_to_parquet_mock: Mock,
@@ -32,9 +52,33 @@ class ImputeIndCqcAscwdsAndPirTests(unittest.TestCase):
         )
 
         scan_parquet_mock.assert_called_once()
+        self.assertEqual(forward_fill_latest_known_value_mock.call_count, 2)
+        calculate_filled_posts_per_bed_ratio_mock.assert_called_once()
+        self.assertEqual(model_roc_trendline_mock.call_count, 2)
+        convert_pir_to_filled_posts_mock.assert_called_once()
+        merge_ascwds_and_pir_filled_post_submissions_mock.assert_called_once()
         self.assertEqual(model_imputation_mock.call_count, 4)
         nullify_ct_values_previous_to_first_submission_mock.assert_called_once()
         sink_to_parquet_mock.assert_called_once_with(
             ANY,
             self.TEST_DESTINATION,
         )
+
+
+class CalculateRollingAverageTests(unittest.TestCase):
+    def test_calculate_rolling_average_returns_expected_values(self):
+        expected_lf = pl.LazyFrame(
+            Data.expected_rolling_average_rows,
+            Schemas.expected_rolling_average_schema,
+            orient="row",
+        )
+        test_lf = expected_lf.drop(IndCQC.posts_rolling_average_model)
+        returned_lf = test_lf.with_columns(
+            job.calculate_rolling_average(
+                column_to_average=IndCQC.ascwds_filled_posts_dedup_clean,
+                period=Data.test_rolling_average_period,
+                columns_to_partition_by=[IndCQC.location_id],
+            ).alias(IndCQC.posts_rolling_average_model)
+        )
+
+        pl_testing.assert_frame_equal(returned_lf, expected_lf)
